requirements.txt
comtypes==1.4.10
cssutils==2.11.1
lxml==5.3.1
Pillow==11.1.0
pyperclip==1.9.0
PySciter==0.6.9
pywin32==308
Requests==2.32.3
selectolax==0.3.28
selenium==4.29.0
